+++ src/extensions/reinitializeOptions.ts
@@ -3,7 +3,7 @@ import {NgSelectOptions} from "../misc";
 
 /**
  * Method reinitialize options for NgSelect
- * @param {NgSelectOptions} options Options to be used for reinitialization
+ * @param options Options to be used for reinitialization
  */
 export function reinitializeOptions<TValue>(options?: NgSelectOptions<TValue>): NgSelectAction<TValue>
 {