+++ src/modules/ngSelectDynamic.module.ts
@@ -1,6 +1,6 @@
 import {NgModule} from '@angular/core';
 
-import {DynamicValueHandlerComponent} from '../plugins/valueHandler';
+import {DynamicValueHandlerComponent} from '../plugins/valueHandler/dynamic/dynamicValueHandler.component';
 
 /**
  * Module for select and its options, allows use of dynamic value handler